[TAN-3876] Moved test setup to shared context
